requirements.txt
torch==2.0.1
torchvision==0.15.2
timm==0.9.2
Pillow==10.0.0
Flask==2.3.2
flask-cors==3.1.1
tqdm==4.66.1